requirements.txt
whisperx
faster-whisper
argparse
numpy
torch
torchvision
torchaudio
transformers
scipy
pyannote.audio
scikit-learn
google-generativeai
python-dotenv
mistralai>=0.4.0
openai